pip-requirements.txt
speedtest-cli
influxdb
